Separate DateChangeTimeKeeper to its own file
